Refactor: border와 borderheader 경로 수정 및 이름 변경
- borderheader가 ArticleListHeader로 이름 변경
