+++ src/components/organisms/category/index.js
@@ -1,4 +1,4 @@
 export { default as Advertisement } from './Advertisement';
-export { default as Board } from '../../pages/articles/ArticleListPage/components/Board';
+export { default as Board } from '../../pages/articles/ArticleListPage/Board';
 export { default as LeftBar } from './LeftBar';
 export { default as PageSelector } from '../../../common/PageSelector';
